+++ sdk/storage/azure-storage-files-datalake/inc/azure/storage/files/datalake/datalake_responses.hpp
@@ -37,7 +37,6 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake { nam
     std::string RequestId;
     std::string ServiceEndpoint;
     std::string Prefix;
-    Azure::Core::Nullable<std::string> PreviousContinuationToken;
     Azure::Core::Nullable<std::string> ContinuationToken;
     std::vector<FileSystemItem> Items;
   }; // struct ListFileSystemsSinglePageResult
@@ -165,7 +164,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake { nam
     std::string RequestId;
   };
 
-  struct GetDataLakePathAccessControlResult
+  struct GetDataLakePathAccessControlListResult
   {
     std::string ETag;
     Core::DateTime LastModified;